+++ snapshot/common/src/main/java/net/xstopho/resourcelibrary/service/platform/IPlatformHelper.java
@@ -1,0 +1,18 @@
+package net.xstopho.resourcelibrary.service.platform;
+
+import java.nio.file.Path;
+
+public interface IPlatformHelper {
+
+    Path getConfigDir();
+
+    default Platforms getPlatform() {
+        return Platforms.NO_LOADER;
+    }
+
+    boolean isModLoaded(String modId);
+
+    enum Platforms {
+        FABRIC, FORGE, NEOFORGE, NO_LOADER
+    }
+}