+++ snapshot/fabric/src/main/java/net/xstopho/resourcelibrary/platform/FabricPlatformHelper.java
@@ -1,0 +1,23 @@
+package net.xstopho.resourcelibrary.platform;
+
+import net.fabricmc.loader.api.FabricLoader;
+import net.xstopho.resourcelibrary.service.platform.IPlatformHelper;
+
+import java.nio.file.Path;
+
+public class FabricPlatformHelper implements IPlatformHelper {
+    @Override
+    public Path getConfigDir() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+
+    @Override
+    public Platforms getPlatform() {
+        return Platforms.FABRIC;
+    }
+
+    @Override
+    public boolean isModLoaded(String modId) {
+        return FabricLoader.getInstance().isModLoaded(modId);
+    }
+}