+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 #### v2.5.0
 - internal cleanup and unification
-- adding a common Event System
+- adding a common Event class
+  - functions like the Fabric Event System